업로드 파일 크기 설정을 킬로에서 바이트로 변경

+++ 2.Modules/checkup/Contracts/function/HDS/HAC/HAC020/featureMain.cs
@@ -277,7 +277,7 @@ TransactionException:
                                         repository.UploadTypeID = uploadTypeID;
                                         repository.UploadExtensions = uploadExtensions;
                                         repository.UploadCount = uploadCount.ParseInt(0);
-                                        repository.UploadSizeLimit = uploadSizeLimit.ParseInt(0);
+                                        repository.UploadSizeLimit = uploadSizeLimit.ParseLong(0);
                                         repository.IsLocalDbFileManaged = isLocalDBFileManaged.ParseBool(false);
                                         repository.SQLiteConnectionString = sqliteConnectionString;
                                         repository.TransactionGetItem = transactionGetItem;
@@ -323,7 +323,7 @@ TransactionException:
                                         repository.UploadTypeID = uploadTypeID;
                                         repository.UploadExtensions = uploadExtensions;
                                         repository.UploadCount = uploadCount.ParseInt(0);
-                                        repository.UploadSizeLimit = uploadSizeLimit.ParseInt(0);
+                                        repository.UploadSizeLimit = uploadSizeLimit.ParseLong(0);
                                         repository.IsLocalDbFileManaged = isLocalDBFileManaged.ParseBool(false);
                                         repository.SQLiteConnectionString = sqliteConnectionString;
                                         repository.TransactionGetItem = transactionGetItem;
